IcyDwarf v13.8.3
Valgrind-clean version.

Changes in the Crack routine:
-Improved treatment of the brittle-ductile transition: switched from the
approach of Mogi (1966) where the transition is just a function of
pressure to a more elaborate model that involves a friction law
(Escartin et al. 1999) and a flow law (Hilairet et al. 2007) for
hydrated rock. The dependence on P, T, and strain rate is included.
-Added a switch for thermal mismatch in the input file as well as for
chemical species to dissolve/precipitate.
-Improved treatment of hydration to shrink/create cracks.

To do:
-Make the Crack routine able to run within a thermal evolution loop.
-In the Cryolava routine, get log K at P_gas locally.