IcyDwarf v14.3.6
IcyDwarfPlot: Structure can now be overlain on the temperature vs.
radius plot. The plot has numbers on the axes using SDL2_ttf, which
needs to be included at compilation.
The thermal view is slow: need to check memory leaks.
Need to replace the graphical numbers on Thermal and crack views with
TTF.

Readme file now had installation and use instructions.