IcyDwarf v14.4.2
IcyDwarf: Improved dehydration routine by setting a threshold of 0.01
for Xhydr under which Xhydr is set to 0. May want to do the same thing
with an upper threshold. The dehydration routine is now called only
where rock is not already dehydrated.

IcyDwarfPlot: Added frame-by-frame functionality. Changed the input file
so it is now possible to specify which thermal file to read. Deleted
spurious inputs specific to IcyDwarf in the input file.