IcyDwarf v14.4.5
This version of IcyDwarf is *NOT* working, nor is it intended to.
For any calculations, use v14.4.4.
First attempt at fully integrating the thermal and cracking codes.

The crack() routine is now called from Thermal() instead of being run
standalone. This has necessitated the modification of
calculate_pressure(), which for now doesn't work.
The code compiles, but there is no guarantee that crack() does what it's
intended to. Benchmarking is in progress.