IcyDwarf v14.11.1
Thermal: Fixed treatment of convection in porous media. The formula for
the Rayleigh number for hydrothermal circulation, taken from Desch et
al. (2009), was valid for fluids but not for porous media. The new,
appropriate formula for Ra is taken from Nield (1997).
Moreover, the value of the thermal conductivity kap1 previously used in
the calculation of Ra was, erroneously, the enhanced effective thermal
conductivity of hydrothermal circulation. It is now the actual thermal
conductivity of the lowest cracked layer.