IcyDwarf v14.11.1
Thermal: Fixed treatment of convection in hydrothermal layers.
The Ra number was previously taken from Desch et al. (2009). This is
appropriate for convection in fluids, but not in porous media. Ra is now
calculated according to Nield (1997).
Moreover, the thermal conductivity kap1 in the Ra calculation was
previously taken as the enhanced effective thermal conductivity due to
convection. Is is now taken as the actual thermal conductivity of the
lowest cracked layer.