IcyDwarf v15.3.1
Thermal(): Modified the subroutine separate() to differentiate an
arbitrary amount of rock, in case a rock mass/volume fraction Xfines is
in small enough grains that it wouldn't settle into a core. For now, the
value of Xfines (between 0 and 1) needs to be hard-wired into the code.

It's unclear whether hydrate() and dehydrate() can be called with an
arbitrary Xfines, so for now the core hydration/dehydration
functionality is deactivated if Xfines is not 0.

Minor update: the ice layer bottom radius, irice, is now updated twice
at each time step instead of only when calling separate(). This ensures
that ammonia is well-mixed, that hydrate() and dehydrate() are dealing
with the correct amounts of water, and that the check of whether there
is enough water to circulate in rock is accurate.