IcyDwarf v15.3.7
IcyDwarfPlot: New ParamExploration_plot() routine to plot results from
the ParamExploration() routine in IcyDwarf, which runs the geochemical
code PHREEQC v3 in batches.
The routine is called from the GUI via a tab at the bottom.

Equilibrium liquid and rock compositions are plotted as pie charts as a
function of redox and pH, water:rock ratios, temperature, and pressure.
For now only species containing potassium are plotted, but more are
coming.