IcyDwarf v15.4.4
What is causing the radiogenic heating to "disappear" is that irice for
convection remains 0 if there is no liquid anywhere, leading to global
convection as an iceball. But if irice is initiated at ircore (as it
should be), the code crashes. An elegant workaround could be to set
irice=ir if (a) there is no liquid and (b) there is ice. For now, in
case there is no liquid, irice is set as ircore and that seems to work.
A variable irice_cv separate from irice was created in order to avoid
any conflicts with subroutines like separate(); this might be what
crashes the code.