IcyDwarf v15.8.18
Thermal() now prevents rebuilding of the internal energy grid when
calling separate(). This is equivalent to assuming that differentiation
proceeds slowly enough for material to thermalize with its surroundings.
Otherwise, the treatment of fines in separate() is such that the
internal energy of rock gets partially averaged between the core and
mantle fines each time separate() is called.