IcyDwarf v16.8.1
Updated PHREEQC core8.dat custom database, now as internally consistent
as possible, with expressions of equilibrium constants as a function of
temperature and pressure for all species. This database is suitable for
equilibrium calculations between 0.01 and 200ºC and 1 and 1000 bar,
limited by the range of validity of the Peng-Robinson equation of state
used for gas species (Appelo, Parkhurst & Post 2014). Without gases, the
database could be safely used up to 300ºC and 5000 bar, the fitting
range of log K=f(T) and Vm for solutes (Johnson, Oelkers & Helgeson
1992).