IcyDwarf v18.2.5
Started coding averaged Hamiltonian routine for determining change in
eccentricity due to moon-moon interactions. The routine implements the
equations described in the Appendix of Meyer & Wisdom (2008). The
routine is not fully coded, so for now the code defaults to the old
forced eccentricity jump routine after Charnoz et al. (2011) (now
accounting for the probability of capture into resonance as of v18.2).
The code should be usable, but hasn't been tested.