IcyDwarf v18.4.5
In the process of coupling the thermal and orbital evolution codes. The
orbital routine is not robust to repeated changes of variables, owing to
slight approximations every time, so the state variables h, k, and a_
will need to be stored if a resonance continues across two thermal time
steps. This version, still in development, holds up to the previous
benchmark.