IcyDwarf v18.5.2
Moon-moon interactions are now handled only outside the influence of the
ring torque, under the assumption that orbital evolution due to the ring
is fast enough that it prevents capture into a moon-moon resonance.
Otherwise, moon-ring interactions should be incorporated into the
da_tide variable in MMR_AvgHam() of Orbit.h.