IcyDwarf v18.5.3
Added in the capability to restart a run from an interrupted simulation.
From benchmarking, the pieced-up run seems to yield similar results if
the object is frozen, but slightly different if the object is just
melted. It remains to be confirmed how much of a departure this is.

In doing so some inefficiencies in the code have been streamlined
(especially in the initialization), let's make sure this doesn't affect
the code's behavior.

On to adding a switch in the input file to choose between normal and
recovery mode.