IcyDwarf v18.6.1
Tweak that will hopefully fix orbits getting in and out of resonance.
Now the condition for moons already in resonance to stay in resonance is
on a__old, rather than the osculating semimajor axes aorb.

Also fixed initialization bug for normal init (missing volumes when
calling separate()) subsequent to the implementation of the recovery
mode.