IcyDwarf v24.7.1
Fixed bug in tail() subroutine of recover() that was preventing reading
of the first value in the output file. However, this fix requires the
output file to have more than the last n lines to be read. That's
satisfied if either more than one timestep have been printed out, or if
the file has been manually padded with, e.g., a line of zeros at the
beginning.